feat: Implement solution for counting the number of teams
This commit adds a new file, `count_number_of_teams.py`, which contains the implementation of a solution for the "Count Number of Teams" problem. The solution uses nested loops to compare the ratings of team members and count the number of valid teams. This implementation improves the readability and performance of the code.

Link to the problem: [Count Number of Teams](https://leetcode.com/problems/count-number-of-teams/?envType=daily-question&envId=2024-07-29)