feat: Implement solution for minimum swaps to group all 1s together
This commit adds a new file, `minimum_swaps_to_group_all_1s_together_ii.py`, which contains the implementation of a solution for the "Minimum Swaps to Group All 1's Together II" problem. The code uses a sliding window approach to find the minimum window that contains the minimum number of 0's, as it requires the least number of swaps to correct. This implementation enhances the efficiency and correctness of the code.

Link to the problem: [Minimum Swaps to Group All 1's Together II](https://leetcode.com/problems/minimum-swaps-to-group-all-1s-together-ii/?envType=daily-question&envId=2024-08-02)